+++ tests/test_exceptions.py
@@ -235,15 +235,6 @@ class ExceptionTest(unittest.TestCase):
         with self.assertRaisesRegex(UBXMessageError, EXPECTED_ERROR):
             UBXMessage("RXM", "RXM-PMP", GET, timeTag=0)
 
-    def testFill_RXMRLMGET(
-        self,
-    ):  #  test RXM-RLM GET constructor without type or payload keyword
-        EXPECTED_ERROR = (
-            "RXM-RLM message definitions must include type or payload keyword"
-        )
-        with self.assertRaisesRegex(UBXMessageError, EXPECTED_ERROR):
-            UBXMessage("RXM", "RXM-RLM", GET, version=0, svId=23)
-
     def testFill_RELPOSNED(
         self,
     ):  #  test NAV_RELPOSNED GET constructor without version or payload keyword
@@ -252,6 +243,15 @@ class ExceptionTest(unittest.TestCase):
         )
         with self.assertRaisesRegex(UBXMessageError, EXPECTED_ERROR):
             UBXMessage("NAV", "NAV-RELPOSNED", GET, relPosN=1, relPosE=2, relPosD=3)
+
+    def testFill_RXMRLMGET(
+        self,
+    ):  #  test RXM-RLM GET constructor without type or payload keyword
+        EXPECTED_ERROR = (
+            "RXM-RLM message definitions must include type or payload keyword"
+        )
+        with self.assertRaisesRegex(UBXMessageError, EXPECTED_ERROR):
+            UBXMessage("RXM", "RXM-RLM", GET, version=0, svId=23)
 
     def testFill_CFGNMEAGET(
         self,